requirements.txt
numpy==1.24.4
casadi==3.6.4
pyyaml==6.0.1
matplotlib==3.7.4
pytest==7.2.0